Update version
This is because the first release workflow run after switching to an unscoped
package name unnecessarily failed due to an error stating that
'https://registry.npmjs.org/procon-ip' was not found while publishing to the
github repo. Actually it was already successfully published to npmjs in that
moment. But maybe not yet reachable via registry api.
Finally in turn the re-run obviously failed in publishing to npmjs, because
the package/version already exists.
I thought about unpublishing the packages and re-run the workflow again. But
this might end in the same situation because the package might be removed
completely from npmjs, after the only existing version has been unplubished.